rename clip.ts to audio-clip.ts
