+++ cocos/audio/audio-clip.ts
@@ -32,9 +32,9 @@ import { AudioPlayer } from 'pal/audio';
 import {
     ccclass, type, serializable, override,
 } from 'cc.decorator';
-import { Asset } from '../../core/assets/asset';
-import { legacyCC } from '../../core/global-exports';
-import { AudioType } from '../../../pal/audio/type';
+import { Asset } from '../core/assets/asset';
+import { legacyCC } from '../core/global-exports';
+import { AudioType } from '../../pal/audio/type';
 
 /**
  * @en
